Added files via upload
I have all the individual files laid out, thus if I ever make changes to one of them, it will be easy to just copy and paste the code instead of downloading the whole project again.

AS OF NOW: There is only a secured login that it not attached to a database. I created a home.jsp but it was purely just a test page and not updated at all.